def come_up_new_results_filename(cls, output_dir, results_method_id, results_method_type_id, +							call_method_id=None, phenotype_method_id=None,\ +							analysis_method_id=None):  		""" +		2010-9-21 +			add argument call_method_id, phenotype_method_id, analysis_method_id +				and incorporate the non-None ones into the filename
replace session.save with session.add

+++ variation/src/Results2DB_250k.py
@@ -203,6 +203,9 @@ class Results2DB_250k(object):
 			sys.exit(4)
 		
 		if output_fname:
+			if os.path.isfile(output_fname):
+				sys.stderr.write("Error: file %s already exists. Skip.\n"%output_fname)
+				return False
 			writer = csv.writer(open(output_fname, 'w'), delimiter='\t')
 		elif cls.marker_pos2snp_id is None:
 			cls.marker_pos2snp_id = cls.get_marker_pos2snp_id(db)
@@ -284,14 +287,14 @@ class Results2DB_250k(object):
 					#construct a new marker
 					marker = SNPs(name=marker_name, chromosome=chr, position=start_pos, end_position=stop_pos, created_by=user)
 					#save it in database to get id
-					session.save(marker)
+					session.add(marker)
 					cls.marker_pos2snp_id[key] = marker	#for the next time to encounter same marker
 					cls.is_new_marker_added = True	#set this flag as new marker was inputted into the dict
 					r = Results(score=score)
 					r.snps = marker
 					del marker
 				r.results_method = rm
-				session.save(r)
+				session.add(r)
 				del r
 			no_of_lines += 1
 		
@@ -303,15 +306,30 @@ class Results2DB_250k(object):
 	
 	submit_results = classmethod(submit_results)
 	
-	def come_up_new_results_filename(cls, output_dir, results_method_id, results_method_type_id):
+	def come_up_new_results_filename(cls, output_dir, results_method_id, results_method_type_id,
+							call_method_id=None, phenotype_method_id=None,\
+							analysis_method_id=None):
 		"""
+		2010-9-21
+			add argument call_method_id, phenotype_method_id, analysis_method_id
+				and incorporate the non-None ones into the filename 
 		2008-05-30
 			to save the filename into db
 		"""
 		output_dir = os.path.join(output_dir, 'type_%s'%results_method_type_id)
 		if not os.path.isdir(output_dir):
 			os.makedirs(output_dir)
-		output_fname = os.path.join(output_dir, '%s_results.tsv'%results_method_id)
+		
+		filename_ls = [results_method_id]
+		if call_method_id is not None:
+			filename_ls.append(call_method_id)
+		if phenotype_method_id is not None:
+			filename_ls.append(phenotype_method_id)
+		if analysis_method_id is not None:
+			filename_ls.append(analysis_method_id)
+		filename_ls = map(str, filename_ls)
+		
+		output_fname = os.path.join(output_dir, '%s_%s_results.tsv'%(results_method_id, '_'.join(filename_ls)))
 		return output_fname
 	
 	come_up_new_results_filename = classmethod(come_up_new_results_filename)
@@ -369,7 +387,7 @@ class Results2DB_250k(object):
 		rmt = session.query(ResultsMethodType).get(results_method_type_id)
 		if not rmt and results_method_type_short_name is not None:	#create a new results method type
 			rmt = ResultsMethodType(short_name=results_method_type_short_name)
-			session.save(rmt)
+			session.add(rmt)
 		
 		if not rmt:
 			sys.stderr.write("No results method type available for results_method_type_id=%s.\n"%results_method_type_id)
@@ -407,7 +425,7 @@ class Results2DB_250k(object):
 		rm.call_method = cm
 		rm.analysis_method = am
 		
-		session.save(rm)
+		session.add(rm)
 		if rmt:
 			rm.results_method_type = rmt
 		
@@ -416,13 +434,15 @@ class Results2DB_250k(object):
 		
 		session.flush()	#not necessary as no immediate query on the new results after this and commit() would execute this.
 		if commit:
-			rm.filename = cls.come_up_new_results_filename(output_dir, rm.id, rm.results_method_type.id)
+			rm.filename = cls.come_up_new_results_filename(output_dir, rm.id, rm.results_method_type.id, \
+										call_method_id=call_method_id, phenotype_method_id=phenotype_method_id,\
+										analysis_method_id=analysis_method_id)
 			if rm.analysis_method_id==13:
 				return_value = cls.copyResultsFile(db, input_fname, rm, user, rm.filename)
 			else:
 				return_value = cls.submit_results(db, input_fname, rm, user, rm.filename)
 			if return_value:
-				session.save_or_update(rm)
+				session.add(rm)
 				# 2010-5-3 store the json structure of top 10000 SNPs from the rm into db
 				no_of_top_snps = 10000
 				if rm.analysis_method.min_maf is not None:
@@ -434,7 +454,7 @@ class Results2DB_250k(object):
 					rm_json = ResultsMethodJson(min_MAF=min_MAF, no_of_top_snps=no_of_top_snps)
 					rm_json.result = rm
 					rm_json.json_data = json_data
-					session.save_or_update(rm_json)
+					session.add(rm_json)
 				except:
 					sys.stderr.write('Except in saving results_method_json (aborted): %s\n'%repr(sys.exc_info()))
 					import traceback
@@ -444,7 +464,6 @@ class Results2DB_250k(object):
 				session.delete(rm)
 			session.flush()
 			session.commit()
-			session.clear()
 			cls.reset_marker_pos2snp_id()
 		else:	#default is also rollback(). to demonstrate good programming
 			session.rollback()
